+++ src/test/java/tools/jackson/databind/struct/UnwrappedWithAnyGetterPrefixTest.java
@@ -11,6 +11,8 @@ import com.fasterxml.jackson.annotation.*;
 import tools.jackson.core.JsonGenerator;
 
 import tools.jackson.databind.*;
+import tools.jackson.databind.annotation.JsonDeserialize;
+import tools.jackson.databind.annotation.JsonPOJOBuilder;
 import tools.jackson.databind.annotation.JsonSerialize;
 import tools.jackson.databind.node.ObjectNode;
 import tools.jackson.databind.ser.std.StdSerializer;
@@ -577,5 +579,89 @@ public class UnwrappedWithAnyGetterPrefixTest extends DatabindTestUtil
     {
         assertNotNull(MAPPER.readValue("""
                 {"name":"x","a-p":1,"b-q":2}""", TwoUnwrappedOuter.class));
+    }
+
+    /*
+    /**********************************************************************
+    /* Test methods: numeric map keys
+    /**********************************************************************
+     */
+
+    @JsonPropertyOrder({ "name" })
+    static class IntKeyOuter {
+        public String name = "aaa";
+
+        @JsonUnwrapped(prefix = "a-")
+        public IntKeyBean inner = new IntKeyBean();
+    }
+
+    static class IntKeyBean {
+        public Map<Integer, Object> extra = new LinkedHashMap<>();
+
+        @JsonAnyGetter
+        public Map<Integer, Object> getExtra() { return extra; }
+    }
+
+    // `Integer`/`Long` keys are written via `JsonGenerator.writePropertyId(long)`,
+    // which needs transforming just like `String`/`Enum` keys do
+    @Test
+    public void intKeyedAnyGetterAppliesPrefix() throws Exception
+    {
+        IntKeyOuter input = new IntKeyOuter();
+        input.inner.extra.put(3, "x");
+
+        assertEquals("""
+                {"name":"aaa","a-3":"x"}""", MAPPER.writeValueAsString(input));
+    }
+
+    /*
+    /**********************************************************************
+    /* Test methods: builder-based unwrapped bean
+    /**********************************************************************
+     */
+
+    @JsonDeserialize(builder = BuilderAnyBean.Builder.class)
+    static class BuilderAnyBean {
+        public final String id;
+        public final Map<String, Object> extra;
+
+        BuilderAnyBean(String id, Map<String, Object> extra) {
+            this.id = id;
+            this.extra = extra;
+        }
+
+        @JsonPOJOBuilder(withPrefix = "")
+        static class Builder {
+            String id;
+            Map<String, Object> extra = new LinkedHashMap<>();
+
+            public Builder id(String id) {
+                this.id = id;
+                return this;
+            }
+
+            @JsonAnySetter
+            public void any(String key, Object value) { extra.put(key, value); }
+
+            public BuilderAnyBean build() { return new BuilderAnyBean(id, extra); }
+        }
+    }
+
+    static class BuilderOuter {
+        public String name;
+
+        @JsonUnwrapped(prefix = "a-")
+        public BuilderAnyBean inner;
+    }
+
+    // Builder-based unwrapped bean: `BuilderBasedDeserializer` has to retain the
+    // `NameTransformer` too, so that the prefix is stripped off any-setter keys
+    @Test
+    public void builderBasedInnerStripsPrefix() throws Exception
+    {
+        BuilderOuter result = MAPPER.readValue("""
+                {"name":"x","a-id":"i1","a-age":64}""", BuilderOuter.class);
+        assertEquals("i1", result.inner.id);
+        assertEquals(Map.of("age", 64), result.inner.extra);
     }
 }
